+++ src/main/java/com/apollo/main/dto/response/ProdutoResponseDTO.java
@@ -8,6 +8,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -29,7 +31,7 @@ public class ProdutoResponseDTO {
     private String nome;
     private String descricao;
     private int qntdEstoque;
-    private Double precoCusto;
-    private Double precoVenda;
+    private BigDecimal precoCusto;
+    private BigDecimal precoVenda;
 
 }
